+++ app/src/main/java/com/techcourse/service/TxUserService.java
@@ -2,7 +2,7 @@ package com.techcourse.service;
 
 import com.techcourse.config.DataSourceConfig;
 import com.techcourse.domain.User;
-import org.springframework.jdbc.core.TransactionManager;
+import org.springframework.transaction.support.TransactionManager;
 
 public class TxUserService implements UserService {
