+++ app/src/test/java/com/techcourse/service/UserServiceTest.java
@@ -12,7 +12,7 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.springframework.dao.DataAccessException;
 import org.springframework.jdbc.core.JdbcTemplate;
-import org.springframework.jdbc.core.TransactionManager;
+import org.springframework.transaction.support.TransactionManager;
 
 class UserServiceTest {
