+++ jdbc/src/main/java/org/springframework/transaction/support/TransactionNoReturnExecutor.java
@@ -1,4 +1,4 @@
-package org.springframework.jdbc.core;
+package org.springframework.transaction.support;
 
 @FunctionalInterface
 public interface TransactionNoReturnExecutor {
